add another case for MPI_Initialized
An additional case was needed when standard MPI code (i.e. with a main()
function without including any boilerplate code) is run under FG-MPI
(nfg defaults to 1 if it's omitted on the command line).
It is also possible to launch a mix of MPI-only code and FG-MPI
code using the colon operator for mpiexec on the command line.
FGP_ALL_POST_INIT now refers to the case when FGmpiexec() is called
in the code for nfg >=1